change int to double on currency fix bugs on add spareparts fix bugs on dashboard and cash flow cannot show data when double is exist

+++ lib/calculator/controller/price_calc_controller.dart
@@ -20,7 +20,7 @@ class PriceCalculatorController extends GetxController {
   //   super.onInit();
   // }
 
-  double calculateFromWidget(int hargaSupplier) {
+  double calculateFromWidget(double hargaSupplier) {
     tambahWaranti = 30;
     tempohWarranti.value = '1 Bulan';
     supplierPrice.value = hargaSupplier.toString();
